feat: add remember session feature (#31)
The "remember session" functionality was added for users. A "rememberMeToken" field was added to the User entity, and the corresponding methods were implemented in the UserRepository.

Additionally, the RememberMeUseCase was created to generate and manage the remember session tokens. The AuthService.ts and options.ts files were also modified to integrate this functionality with NextAuth.

closes #2